+++ sync/SyncDataCollection.ts
@@ -29,6 +29,12 @@ class SyncDataCollection {
     private getLastSyncDownTimestamp: (table: DataCollection<any, any>) => number;
     /** Set the last synched timestamp of a particular table */
     private updateLastSyncDownTimestamp: (table: DataCollection<any, any>) => void;
+    /** An optional action start function */
+    private notifyActionStart: (action: SyncDataCollection.SyncAction, table: DataCollection<any, any>) => any;
+    /** An optional action end function */
+    private notifyActionEnd: (action: SyncDataCollection.SyncAction, table: DataCollection<any, any>, startTimerKey: any) => void;
+    /** An optional action failed function */
+    private notifyActionFailure: (action: SyncDataCollection.SyncAction, table: DataCollection<any, any>, startTimerKey: any, err: any) => void;
 
 
     /** Create an object which can sync a data collection to/from a remote data destination/source
@@ -37,14 +43,24 @@ class SyncDataCollection {
      * @param isDeletedPropName the name of the property on both local and remote data models which contains the item's 'deleted' boolean flag
      * @param isSynchedPropName the name of the property on both local and remote data models which contains the item's 'synched' boolean flag
      * @param lastModifiedPropName the name of the property on both local and remote data models which contains the item's last-modified unix style millisecond timestamp number
+     * @param [notifyActionStart] optional event listener type function which is called whenever a sync action starts, see {@link SyncAction}
+     * @param [notifyActionEnd] optional event listener type function which is called whenever a sync action finishes, see {@link SyncAction}. Note: this method is only called if the action is successful, see 'notifyActionFailure'
+     * @param [notifyActionFailure] optional event listener type function which is called whenever a sync action fails, see {@link SyncAction}
      */
     constructor(getLastSyncDownTimestamp: (table: DataCollection<any, any>) => number, updateLastSyncDownTimestamp: (table: DataCollection<any, any>) => void,
-            isDeletedPropName: string, isSynchedPropName: string, lastModifiedPropName: string) {
+        isDeletedPropName: string, isSynchedPropName: string, lastModifiedPropName: string,
+        notifyActionStart?: (action: SyncDataCollection.SyncAction, table: DataCollection<any, any>) => any,
+        notifyActionEnd?: (action: SyncDataCollection.SyncAction, table: DataCollection<any, any>, startTimerKey: any) => void,
+        notifyActionFailure?: (action: SyncDataCollection.SyncAction, table: DataCollection<any, any>, startTimerKey: any, err: any) => void
+    ) {
         this.getLastSyncDownTimestamp = getLastSyncDownTimestamp;
         this.updateLastSyncDownTimestamp = updateLastSyncDownTimestamp;
         this.isDeletedPropName = isDeletedPropName;
         this.isSynchedPropName = isSynchedPropName;
         this.lastModifiedPropName = lastModifiedPropName;
+        this.notifyActionStart = notifyActionStart;
+        this.notifyActionEnd = notifyActionEnd;
+        this.notifyActionFailure = notifyActionFailure;
     }
 
 
@@ -79,7 +95,7 @@ class SyncDataCollection {
      * @param <R2> the process results callback error type
      */
     public syncDownCollection<E, F, P, S, R1, R2>(params: P, table: DataCollection<E, F>, syncDownFunc: (params: P) => PsPromise<S[], R1>,
-            processResultsCallback: (items: S[]) => void | Q.IPromise<R2>): PsPromise<void, SyncError> {
+            processResultsCallback: (items: S[]) => void | PsPromise<void, R2>): PsPromise<void, SyncError> {
         var self = this;
         var dfd = Defer.newDefer<void, SyncError>();
 
@@ -89,12 +105,25 @@ class SyncDataCollection {
                 syncingDown: true,
                 error: err,
             });
+
+            if (self.notifyActionFailure) {
+                if (!isAfterSync) {
+                    self.notifyActionFailure("syncDown", table, syncDownTimer, err);
+                }
+                else {
+                    self.notifyActionFailure("afterSyncDownUpdate", table, afterSyncDownUpdateTimer, err);
+                }
+            }
         }
 
         function saveData() {
             // update the last sync time for this table to right now
             try {
                 self.updateLastSyncDownTimestamp(table);
+
+                if (self.notifyActionEnd) {
+                    self.notifyActionEnd("afterSyncDownUpdate", table, afterSyncDownUpdateTimer);
+                }
             } catch (err) {
                 syncFailure(err);
                 return;
@@ -103,12 +132,22 @@ class SyncDataCollection {
         }
 
         try {
+            var syncDownTimer = this.notifyActionStart ? this.notifyActionStart("syncDown", table) : null;
+            var isAfterSync = false;
+            var afterSyncDownUpdateTimer = null;
+
             syncDownFunc(params).done(function (items) {
                 try {
+                    if (self.notifyActionEnd) {
+                        self.notifyActionEnd("syncDown", table, syncDownTimer);
+                    }
+                    isAfterSync = true;
+                    afterSyncDownUpdateTimer = self.notifyActionStart ? self.notifyActionStart("afterSyncDownUpdate", table) : null;
+
                     var promise = processResultsCallback(items);
 
                     if (promise != null && promise["then"]) {
-                        (<Q.IPromise<R2>>promise).then(saveData, syncFailure);
+                        (<PsPromise<void, R2>>promise).then(saveData, syncFailure);
                     } else {
                         saveData();
                     }
@@ -132,25 +171,42 @@ class SyncDataCollection {
      * only contains one record and send that one record as an object, rather than sending an
      * array of objects to the service call, false or undefined sends an array of any data in the collection
      */
-    public syncUpCollection<E, F, P, S, U, R>(params: P, syncSetting: SyncSettingsWithUp<E, F, P, S, U, R>, copyItemFunc: (item: E) => E): PsPromise<U, SyncError> {
+    public syncUpCollection<E, F, P, S, U, R>(params: P, syncSetting: SyncSettingsWithUp<E, F, P, S, U, R>): PsPromise<U, SyncError> {
         var self = this;
         var primaryKeys = syncSetting.primaryKeys;
         var primaryKey = Arrays.getIfOneItem(primaryKeys);
+        var primaryKeyCheckers = syncSetting.hasPrimaryKeyCheckers;
+        var primaryKeyChecker = Arrays.getIfOneItem(primaryKeyCheckers);
         var localColl = syncSetting.localCollection;
 
-        return this.syncAndUpdateCollection(localColl, copyItemFunc, primaryKey, primaryKeys, function convertAndSendItemsToServer(items) {
+        return this.syncUpAndUpdateCollection(localColl, primaryKey, primaryKeys, function convertAndSendItemsToServer(items) {
+            var beforeSyncUpPrepTimer = self.notifyActionStart ? self.notifyActionStart("beforeSyncUpPrep", localColl) : null;
+
             var toSvcObj = syncSetting.convertToSvcObjectFunc;
             var data = null;
             if (primaryKey) {
-                data = SyncDataCollection.checkAndConvertSingleKeyItems(localColl.getName(), items, primaryKey, toSvcObj);
+                data = SyncDataCollection.checkAndConvertSingleKeyItems(localColl.getName(), items, primaryKeyChecker, toSvcObj);
             }
             else {
-                data = SyncDataCollection.checkAndConvertMultiKeyItems(localColl.getName(), items, primaryKeys, toSvcObj);
+                data = SyncDataCollection.checkAndConvertMultiKeyItems(localColl.getName(), items, primaryKeyCheckers, toSvcObj);
             }
 
+            if (self.notifyActionEnd) {
+                self.notifyActionEnd("beforeSyncUpPrep", localColl, beforeSyncUpPrepTimer);
+            }
+            var syncUpTimer = self.notifyActionStart ? self.notifyActionStart("syncUp", localColl) : null;
+
             return syncSetting.syncUpFunc(params, data).then(function (res) {
+                if (self.notifyActionEnd) {
+                    self.notifyActionEnd("syncUp", localColl, syncUpTimer);
+                }
+
                 return res;
             }, function (err): Throws<SyncError> {
+                if (self.notifyActionFailure) {
+                    self.notifyActionFailure("syncUp", localColl, syncUpTimer, err);
+                }
+
                 throw {
                     collectionName: localColl.getName(),
                     syncingUp: true,
@@ -173,7 +229,7 @@ class SyncDataCollection {
      * @param primaryKeys the table data model's primary keys, this or 'primaryKey' must not be null
      * @param syncAction the action which performs the data sync
      */
-    public syncAndUpdateCollection<E, F, R, S>(table: DataCollection<E, F>, copyItemFunc: (item: E) => E, primaryKey: string, primaryKeys: string[], syncAction: (items: E[]) => PsPromise<R, S>): PsPromise<R, S> {
+    public syncUpAndUpdateCollection<E, F, R, S>(table: DataCollection<E, F>, primaryKey: string, primaryKeys: string[], syncAction: (items: E[]) => PsPromise<R, S>): PsPromise<R, S> {
         var self = this;
         var dfd = Defer.newDefer<R, S>();
 
@@ -187,15 +243,20 @@ class SyncDataCollection {
             return dfd.promise;
         }
 
-        var itemsData = items.map(copyItemFunc);
+        syncAction(items).done(function (res) {
+            var afterSyncUpUpdateTimer = self.notifyActionStart ? self.notifyActionStart("afterSyncUpUpdate", table) : null;
 
-        syncAction(itemsData).done(function (res) {
             if (primaryKey) {
                 self.updateSinglePrimaryKeyItems(table, items, primaryKey);
             }
             else {
                 self.updateMultiPrimaryKeyItems(table, items, primaryKeys);
             }
+
+            if (self.notifyActionEnd) {
+                self.notifyActionEnd("afterSyncUpUpdate", table, afterSyncUpUpdateTimer);
+            }
+
             dfd.resolve(res);
         }, function (err) {
             dfd.reject(err);
@@ -254,7 +315,7 @@ class SyncDataCollection {
      * @param isDeletedPropName the name of the property which indicates if an item should be deleted, if value of this property is truthy the item is deleted when the syncing, nullable
      * @param syncDownOp the type of merge/update/add operation to perform with the new items
      */
-    private static createAddUpdateOrRemoveItemsFunc<E, F, P, S, R>(syncSettings: SyncSettingsWithDown<E, F, P, S, R>, isDeletedPropName: string, syncDownOp: SyncDataCollection.SyncDownOp) {
+    public static createAddUpdateOrRemoveItemsFunc<E, F, P, S, R>(syncSettings: SyncSettingsWithDown<E, F, P, S, R>, isDeletedPropName: string, syncDownOp: SyncDataCollection.SyncDownOp) {
         return function addUpdateOrRemoveItemsFunc(items: S[]) {
             var table = syncSettings.localCollection;
             var findFilterFunc = syncSettings.findFilterFunc;
@@ -312,14 +373,15 @@ class SyncDataCollection {
      * Else convert the item using the provided conversion function
      * @return the 'items' array converted to result objects
      */
-    private static checkAndConvertMultiKeyItems<T, R>(collName: string, items: T[], primaryKeyFields: string[], itemConverter: (obj: T) => R): R[] {
-        var keyCount = primaryKeyFields.length;
+    private static checkAndConvertMultiKeyItems<T, R>(collName: string, items: T[], hasPrimaryKeyFuncs: ((obj: T) => boolean)[], itemConverter: (obj: T) => R): R[] {
+        var keyCount = hasPrimaryKeyFuncs.length;
         var resultItems: R[] = [];
         for (var i = 0, size = items.length; i < size; i++) {
             var item = items[i];
             var hasPrimaryKeys = true;
             for (var k = 0; k < keyCount; k++) {
-                if (!item[primaryKeyFields[k]]) {
+                var hasPrimaryKey = hasPrimaryKeyFuncs[k](item);
+                if (!hasPrimaryKey) {
                     hasPrimaryKeys = false;
                     break;
                 }
@@ -339,11 +401,11 @@ class SyncDataCollection {
      * Else convert the item using the provided conversion function
      * @return the 'items' array converted to result objects
      */
-    private static checkAndConvertSingleKeyItems<T, R>(collName: string, items: T[], primaryKeyField: string, itemConverter: (obj: T) => R): R[] {
+    private static checkAndConvertSingleKeyItems<T, R>(collName: string, items: T[], hasPrimaryKeyFunc: (obj: T) => boolean, itemConverter: (obj: T) => R): R[] {
         var resultItems: R[] = [];
         for (var i = 0, size = items.length; i < size; i++) {
             var item = items[i];
-            var hasPrimaryKey = !!item[primaryKeyField];
+            var hasPrimaryKey = hasPrimaryKeyFunc(item);
             if (hasPrimaryKey) {
                 resultItems.push(itemConverter(item));
             }
@@ -357,6 +419,12 @@ class SyncDataCollection {
 }
 
 module SyncDataCollection {
+
+    /** Names of the various sync action event which can occur.
+     * 'before' and 'after' events involve data processing, queries, updating local data collections, etc.
+     * 'syncDown' and 'syncUp' are the actual action calls.
+     */
+    export type SyncAction = ("syncDown" | "syncUp" /*| "beforeSyncDownPrep"*/ | "beforeSyncUpPrep" | "afterSyncDownUpdate" | "afterSyncUpUpdate");
 
 
     /** Definitions of how to sync down data and merge it with local data, currently includes: