+++ sync/syncing-types.d.ts
@@ -20,6 +20,7 @@ declare interface SyncError {
 declare interface SyncSettings<E, F, S, R> {
     localCollection: DataCollection<E, F>;
     primaryKeys: string[];
+    hasPrimaryKeyCheckers: ((obj: E) => boolean)[];
     findFilterFunc: (item: S) => F;
     copyObjectFunc: (item: E) => E;
     convertUrlToSyncDownFunc?: (url: string) => (params: any) => PsPromise<S[], R>;